NP-2230 Validate NPI subject
