+++ src/pages/registration/references_tab/sub_type_forms/BookForm.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components';
 import { Checkbox, FormControlLabel, MuiThemeProvider, Typography } from '@material-ui/core';
 import BackgroundDiv from '../../../../components/BackgroundDiv';
 import lightTheme from '../../../../themes/lightTheme';
-import { BookType, DescriptionFieldNames, ReferenceFieldNames } from '../../../../types/publicationFieldNames';
+import { BookType, ReferenceFieldNames } from '../../../../types/publicationFieldNames';
 import { BookRegistration } from '../../../../types/registration.types';
 import DoiField from '../components/DoiField';
 import IsbnListField from '../components/IsbnListField';
@@ -14,7 +14,7 @@ import PeerReview from '../components/PeerReview';
 import PublisherField from '../components/PublisherField';
 import SeriesField from '../components/SeriesField';
 import TotalPagesField from '../components/TotalPagesField';
-import DisciplineSearch from '../../description_tab/DisciplineSearch';
+import { NpiDisciplineField } from '../components/NpiDisciplineField';
 
 const StyledSection = styled.div`
   display: grid;
@@ -37,7 +37,7 @@ const StyledTextBook = styled.div`
 
 const BookForm = () => {
   const { t } = useTranslation('registration');
-  const { values, setFieldValue } = useFormikContext<BookRegistration>();
+  const { values } = useFormikContext<BookRegistration>();
   const {
     reference: {
       publicationContext,
@@ -51,18 +51,9 @@ const BookForm = () => {
         <DoiField />
         <PublisherField />
 
-        <Field name={DescriptionFieldNames.NPI_SUBJECT_HEADING}>
-          {({ field: { name, value } }: FieldProps<string>) => (
-            <MuiThemeProvider theme={lightTheme}>
-              <DisciplineSearch
-                setValueFunction={(npiDiscipline) => setFieldValue(name, npiDiscipline?.id ?? '')}
-                dataTestId="search_npi"
-                value={value}
-                placeholder={t('description.search_for_npi_discipline')}
-              />
-            </MuiThemeProvider>
-          )}
-        </Field>
+        <MuiThemeProvider theme={lightTheme}>
+          <NpiDisciplineField />
+        </MuiThemeProvider>
 
         <StyledSection>
           <IsbnListField />
